Rename package window to windows
