+++ src/main/java/com/epam/gene_analyzer/windows/JPopMenuListener.java
@@ -1,4 +1,4 @@
-package com.epam.gene_analyzer.window;
+package com.epam.gene_analyzer.windows;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
